Created a Chaser class to follow player within a certain range. Mob class entities can now move in different speed. collision detection has been improved. created 2_player_sheet.png and 3_player_sheet.png to figure out the outline and center of the player. created Node and Vertor classes. both are going to be used for the A* algorithm path finding. Preparation has been made.

+++ src/com/icyfillup/rain/entity/projectile/WizardProjectile.java
@@ -8,7 +8,7 @@ public class WizardProjectile extends Projectile
 {
 	public static final int FIRE_RATE = 10; // higher is slower
 	
-	public WizardProjectile(int x, int y, double dir)
+	public WizardProjectile(double x, double y, double dir)
 	{
 		super(x, y, dir);
 		range = 200;
